requirements.txt
torch>=2.0
torchvision>=0.15
tqdm>=4.65
pyyaml>=6.0
